Reintroduce PHP 8.1-8.3 support
