+++ .github/workflows/continuous-integration.yml
@@ -15,6 +15,9 @@ jobs:
         operating-system:
           - ubuntu-latest
         php-version:
+          - '8.1'
+          - '8.2'
+          - '8.3'
           - '8.4'
     steps:
       - name: Checkout
