+++ composer.json
@@ -22,7 +22,7 @@
         }
     ],
     "require": {
-        "php": "^8.4",
+        "php": ">=8.1",
         "psr/log": "^1.0 || ^2.0 || ^3.0",
         "ext-pdo": "*"
     },
